* coders/pnm.c (ReadPNMImage): Fix multi-thread issue detected by valgrind's helgrind tool.  Diminish compilation warnings.
* coders/dpx.c (ReadDPXImage): Diminish compilation warnings.

* magick/random.c (AcquireMagickRandomKernel): Fix potential
multi-thread issue detected by valgrind's helgrind tool.

* magick/magick.c (InitializeMagick): Semaphore subsystem needs to be
initialized before anything which uses it.

* magick/semaphore.c (InitializeSemaphore): Since we are using
PTHREAD_MUTEX_INITIALIZER to initialize primary POSIX mutex in the
semaphore subsystem, we should not explicitly initialize the
semaphore a second time.

* magick/segment.c (Classify): Fix multi-thread issue detected by
valgrind's helgrind tool.

* magick/render.c (DrawAffineImage): Use InterpolateViewColor() to
evalute a bi-linear interpolated point rather than obtaining a
pixel value from a close pixel.  This provides better results.

+++ magick/blob.c
@@ -4320,7 +4320,7 @@ MagickExport size_t WriteBlobByte(Image *image,const magick_uint8_t value)
     case StandardStream:
     case PipeStream:
       {
-        if(putc(value,image->blob->file) != EOF)
+        if(putc((int) value,image->blob->file) != EOF)
           return 1;
         return 0;
       }
